BlackHole documentation

BlackHole documentation

Structures

Functions

Internal Methods

Utility Methods

Rendering Kernels

Structures

BlackHole.EddingtonFinkelstein — Type

Eddington-Finkelstein metric describing a Black Hole in spherically symmetric space.

BlackHole.GaussianThinDisk — Type

Disk with Gaussian profile for opacity.

BlackHole.GeometricDisk — Type

Geometric disk with no opacity. Does not consider relativistic, doppler or any other effects.

BlackHole.Singularity — Type

Super type of all space time metrics.

Functions

BlackHole.calcgeodesics — Function

function calcgeodesics(
    s::Singularity;
    num::Int=1500,
    Δϕ::Float64=0.004,
)::Array{Float64,3}

Calculate integrated geodesic curves for a given singularity.

Performs num geodesic traces, each differing in initial velocity 4-vector by an angle Δϕ in the x, y plane.

Note, the method for extracting the results from the integration solutions, and the actual method of integration, is quite tentative at the moment, and requires some investigation.

If num is too low, it can lead to aliasing problems in the final render.

If Δϕ is too low, the integrator calculates light paths which swing around and terminate in odd locations, leading to "jets" streaming out of the black hole into the plane of the disk (where there really should not be any jets).

Extended help

The root cause of the integration issue lies probably in how the integration chart was adapted with the observer function. A more discriminating observer may improve the renderings significantly.

BlackHole.renderdisk — Function

renderdisk(
    ::Val{:gpu},
    disk::AccretionDisk,
    geodesics::AbstractArray{<:Number},
    width::Int,
    height::Int,
    fov::Int,
)

Dispatch method for rendering the disk using the GPU.

Extended help

The kernel launch configuration is hard-coded into this function, which launches two kernels: kernel_index_calculator and kernel_geodesic_render. See the documentation of these individual functions for more details.

renderdisk(
    ::Val{:cpu},
    disk::AccretionDisk,
    geodesics::AbstractArray{<:Number},
    width::Int,
    height::Int,
    fov::Int,
)

Dispatch method for rendering the disk using the CPU.

renderdisk(
    disk::AccretionDisk,
    geodesics::AbstractArray{<:Number};
    width::Int = 480,
    height::Int = 720,
    fov::Int = 200
)

Render an accretion disk disk given an array of pre-calculated geodesics geodesics for the disk.

Returns an array of dimensions width by height, representing the rendered image.

This method checks whether CUDA is installed, and dispatches a GPU kernel accordingly. Otherwise, the rendering functions are executed in parallel on the CPU.