Internal Methods

Documentation for methods not exported by the module, included for development purposes.

Utility Methods

BlackHole.dspher2xyz — Method

dspher2xyz(r, θ, ϕ, dr, dθ, dϕ, disk::AccretionDisk)

Transform differential coordinates from spherical to cartesian, at a given point r, θ, ϕ, also performing relevant transformation for plane of disk.

Returns (dx, dy, dz).

Takes Jacobian into account. Symbolic expressions originally calculated with SageMaths.

Complement to dxyz2spher.

BlackHole.dxyz2spher — Method

dxyz2spher(x, y, z, dx, dy, dz)

Transform differential coordinates from cartesian to spherical, at a given point x, y, z.

Returns (dr, dθ, dϕ).

Takes Jacobian into account. Symbolic expressions originally calculated with SageMaths.

Complement to dspher2xyz.

BlackHole.inclination_transform — Method

inclination_transform(x, y, z, α, β)

Transform a point on x,y,z to the inclination given by a rotation of α around x, and β around z.

Returns (x, y, z) in the rotated plane.

BlackHole.spher2xyz — Method

spher2xyz(t, r, θ, ϕ)

Transform coordinates of a space time point in spherical to cartesian coordinates.

Returns (t, x, y, z).

BlackHole.truncator — Method

truncator(px)::UInt8

Clamps value of px between 0 and 255.

Rendering Kernels

BlackHole.kernel_geodesic_render — Function

kernel_geodesic_render(
    indeximage::AbstractArray{T},
    β_store::AbstractArray{<:Number},
    geo_matrix,::AbstractArray{<:Number},
    width::Int, height::Int,
    step_num::Int,
    disk::GeometricDisk,
    max_index::Int
) where T <: Int

CUDA kernel for computing the integrated paths of geodesics through an accretion disk. indeximage is the result from calling kernel_index_calculator, and is a matrix containing the geodesic index for each pixel.

max_index is the maximal index of the geodesics matrix, i.e. size(geo_matrix, 3).

To be launch with threads = 512 blocks = length(indeximage), until better topology is implemented.

Extended help

This kernel works by using one compute block per pixel, and calculating the difference between two points in spacetime along the geodesic per thread. The thread then calculates whether the line length between these points intersects the accretion disk, and sets a value in a shared memory buffer: in essense, each thread calculates

    geodesics[:, thread + 1, geoindex] - geodesics[:, thread, geoindex]

transformed by the angles of the accretion disk. This allows the intersection to be calculated from the $z$ component: $0 \leq \frac{z_i}{z_i - z_{i+1}} < 1$

Internally, the offsets are mapped with the column major formula

n x m x p:
    arr[i, j, k] == arr[(n*m)*(k-1) + n*(j-1) + i]

to avoid getindex function calls.

At the end of the kernel, a simple reduce algorithm is implemented to sum over the shared array. The first thread of each block then stores the value back into indeximage as the output of the kernel.

Better topology for this kernel would have the y dimension continue to calculate the same pixel, incase there are more geodesic integration steps (step_num) than threads can be spawned. Due to the size of this kernel, only about 512 threads can be spawned on my hardware before the device runs out of resources.

BlackHole.kernel_index_calculator — Function

function kernel_index_calculator(
    oimg::AbstractArray{Int},
    β_store::AbstractArray{<:Number},
    width::Int,
    height::Int,
    fov::Int,
    num_geodesics::Int
)

CUDA kernel for calculating which index of the geodesic array to use in each pixel. The index calculated for a given y, x position is stored in oimg[y, x]. β_store[y, x] also stores the corresponding pixel's angle around the x axis, used for calculating inclination into the accretion disk in the second kernel (kernel_geodesic_render).

fov is a "field of view" index, used to adjust which pixel maps to the end of the geodesic array (num_geodesics). As a rough guideline, fov=1000 maps the pixel at distance r=1000 to the last index of the geodesic array.